refactor(phase4c): move collision.ts, fill.ts to features/grid-editor/utils/
Move grid-specific utilities from src/utils/ to src/features/grid-editor/utils/
as part of Phase 4 folder structure redesign.

Utils moved:
- collision.ts (collision detection, blocked zones, layer calculations)
- fill.ts (fillAllWithSize, fillGaps, getLayerCoverage)

Changes:
- Create utils/ directory in features/grid-editor/
- Move utils with updated imports
- Create re-export stubs at old locations
- Add barrel export index.ts
